backend/pyproject.toml
[tool.poetry]
name = "nostr-relay-explorer"
version = "0.0.1"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9.13"
requests = "^2.28.1"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.14.0"
uwsgi = "^2.0.21"
psycopg2 = "^2.9.5"
django-rest-passwordreset = "^1.3.0"
django = "^4.1"
huey = "^2.4.5"
redis = "^4.6.0"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
ipdb = "^0.13.9"
pytest = "^7.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
extend_skip = ["__init__.py"]
extend_skip_glob = ["*/migrations/*"]

[tool.black]
extend-exclude = '''
/(
  | migrations
)/
'''
